backend/requirements.txt
fastapi
uvicorn
boto3
python-multipart
python-dotenv
pytest
httpx
pytest-asyncio